+++ Game/enemy/property.c
@@ -1,4 +1,7 @@
 #include "property.h"
+#include "math.h"
+
+double square(double x);
 
 void Enemy_Init(struct Enemy *enemy,struct Object* body,struct Object* face, struct Object* leg){
     enemy->body = *body;
@@ -20,13 +23,41 @@ void Enemy_Init(struct Enemy *enemy,struct Object* body,struct Object* face, str
 	CollideBox_Set(&enemy->collideBox, &v1, &v2);
 }
 
-void Del_Enemy(struct Enemy *enemy)
-{
-	free(enemy);
-}
-
 void Enemy_Start(struct Enemy *enemy)
 {
 	//TODO enemy start
 }
 
+bool Damage_Determination(struct Scene *scene,struct Enemy *enemy)//enemy's damage to player
+{
+    double damage = 0;
+    double distance = 1.5;//max distance for enemy to attack the player
+    if (sqrt(square(enemy->transform.globalPosition.x - scene->player.transform.globalPosition.x) +
+             square(enemy->transform.globalPosition.z - scene->player.transform.globalPosition.z)
+             >= distance)) return 0;
+    return 1;
+}
+
+double Damage_Calculation(struct Scene *scene,struct Enemy *enemy)//enemy's damage to player
+{
+    double roller = rand() % 100;
+    double rate = roller * enemy->Critical_Rate;//this value should be 0-5000
+    if (roller >= 2345/*critical damage*/){ return square(enemy->damage/scene->player.defence); }
+    else { return enemy->damage/scene->player.defence; }
+}
+
+void Enemy_Move(struct Enemy* enemy, struct Vector3* move){
+    Transform_AddPosition(&enemy->transform, move);
+}
+
+void Enemy_Rotation(struct Enemy* enemy, struct Vector3 *angle){
+    Object_Rotation(&enemy->face,angle);
+    Object_Rotation(&enemy->body,angle);
+    Vector3_Add(&enemy->transform.rotation, angle);
+    Transform_RotationMatrixUpdate(&enemy->transform);
+}
+
+double square(double x)
+{
+    return x*x;
+}
