+++ Game/enemy/property.h
@@ -3,6 +3,7 @@
 
 #include "../../op_engine/op_engine.h"
 #include "../player.h"
+#include "../scene.h"
 
 
 struct Enemy{
@@ -15,10 +16,13 @@ struct Enemy{
 
 void Enemy_Init(struct Enemy *enemy,struct Object* body,struct Object* face, struct Object* leg);
 
-void Del_Enemy(struct Enemy *enemy);
-
 void Enemy_Start(struct Enemy *enemy);
 
+bool Damage_Determination(struct Scene *scene,struct Enemy *enemy);
 
+double Damage_Calculation(struct Scene *scene,struct Enemy *enemy);
 
+void Enemy_Move(struct Enemy* enemy, struct Vector3 *move); //enemy consists of three objects: face and body
+
+void Enemy_Rotation(struct Enemy *enemy, struct Vector3* angle);
 #endif //FPS_SIMULATOR_PROPERTY_H
