+++ Game/player.h
@@ -1,6 +1,5 @@
 #ifndef FPS_SIMULATOR_PLAYER_H
 #define FPS_SIMULATOR_PLAYER_H
-
 #include "../op_engine/op_engine.h"
 
 struct Weapon{
@@ -47,4 +46,10 @@ void Player_RotateRight(struct Player *player);
 
 void Player_Control(struct Player *player);
 
+//Initializer of weapon (corresponding to player)
+void Weapon_Init(struct Weapon* weapon, struct Player* player, struct ArrayList* weapon_list);
+
+void Weapon_Change(struct Weapon* Old, struct Weapon* New, struct Player* player, struct ArrayList* weapon_list);
+
+void WeaponList_Init(struct ArrayList* weapon_list);
 #endif //FPS_SIMULATOR_PLAYER_H
