+++ Game/scene.c
@@ -143,10 +143,10 @@ void Scene_EnemyCollided(struct Scene *scene, struct Line *ray, struct Enemy **r
 
 void Enemy_Move(struct Enemy* enemy, struct Vector3* move);
 
-void Enemy_Update(struct Enemy* enemy, bool do_find_way, struct Scene* scene, struct link* current)
+void Scene_EnemyUpdate(struct Enemy* enemy, bool do_find_way, struct Scene* scene, struct Enemy_TransformLink* current)
 {
     if (do_find_way){
-        struct link* beginning = Find_Way(&enemy->body,scene);//beginning actually refers to current position of bot.
+        struct Enemy_TransformLink* beginning = Enemy_FindWay(&enemy->body,scene);//beginning actually refers to current position of bot.
         current = beginning;
     }
     struct Transform* target = current->next->current;
@@ -159,13 +159,9 @@ void Enemy_Update(struct Enemy* enemy, bool do_find_way, struct Scene* scene, st
     Enemy_Move(enemy, move);
 }
 
-void Enemy_Move(struct Enemy* enemy, struct Vector3* move){
-    Transform_AddPosition(&enemy->transform, move);
-}
-
 #define typename struct Object
 //build a linked list to find the shortest way from enemy to player.
-struct link* Find_Way(struct Object *enemy, struct Scene* scene)
+struct Enemy_TransformLink* Enemy_FindWay(struct Object *enemy, struct Scene* scene)
 {
     unsigned int size = scene->list_Object.size;
     double min_distance = INFINITY, distance;
@@ -177,10 +173,10 @@ struct link* Find_Way(struct Object *enemy, struct Scene* scene)
     min_distance_obj->transform.globalPosition.x = player->transform.globalPosition.x;
     min_distance_obj->transform.globalPosition.y = player->transform.globalPosition.y;
     min_distance_obj->transform.globalPosition.z = player->transform.globalPosition.z;
-    struct link* destination = NULL;
+    struct Enemy_TransformLink* destination = NULL;
     destination->next = NULL;
     destination->current = &min_distance_obj->transform;
-    struct link* next_addr = destination;
+    struct Enemy_TransformLink* next_addr = destination;
 
     while(!	(min_distance_obj->transform.globalPosition.x == enemy->transform.globalPosition.x &&
                 min_distance_obj->transform.globalPosition.z == enemy->transform.globalPosition.z)){
@@ -202,7 +198,7 @@ struct link* Find_Way(struct Object *enemy, struct Scene* scene)
                 min_temp = &list[j];
             }
         }
-        struct link* link = NULL;
+        struct Enemy_TransformLink* link = NULL;
         link->current = &min_temp->transform;
         link->next = next_addr;
         min_distance_obj = min_temp;
@@ -210,7 +206,7 @@ struct link* Find_Way(struct Object *enemy, struct Scene* scene)
         min_temp = NULL;
     }
 
-    struct link* beginning = NULL;
+    struct Enemy_TransformLink* beginning = NULL;
     beginning->current = &enemy->transform;
     beginning->next = next_addr;
